+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Prefix tool name with server to LLM: <server>__<toolname>. #196
 - Remove `eca_` prefix from eca tools, we already pass server prefix (eca) after #196.
 - Add `approval` arg to preToolCall hook input.
+- Add error rewrite message content message.
 
 ## 0.77.1